Score PopUp Canvas prefab
This ScorePopUpCanvas prefab needs to be put as a child to the desired game object that wants to have the score pop up message on it. And positioned as desired of course.

(And at that point it might be wise to unpack the prefab so that we don't use nested prefabs for example.)